Revert "Update Last Location. Solution for #1169 (#1453)"
This reverts commit d12ff6c700e003202fd2e9842499324561246b15.

The reverted commit was not ready for merging, as it caused dozens of
test failures (around 60), and appeared to fundamentally break parsing
of many stored locations by causing locations to be parsed with
HostAndPort.fromString(location) that were not created with
HostAndPort.toString(). This may be a more fundamental problem with some
of the code in Ample; further investigation and testing is needed before
merging.

+++ server/base/src/main/java/org/apache/accumulo/server/master/state/ZooTabletStateStore.java
@@ -103,7 +103,13 @@ class ZooTabletStateStore implements TabletStateStore {
   }
 
   @Override
-  public void setFutureLocation(Assignment assignment) {
+  public void setFutureLocations(Collection<Assignment> assignments)
+      throws DistributedStoreException {
+    if (assignments.size() != 1)
+      throw new IllegalArgumentException("There is only one root tablet");
+    Assignment assignment = assignments.iterator().next();
+    if (assignment.tablet.compareTo(RootTable.EXTENT) != 0)
+      throw new IllegalArgumentException("You can only store the root tablet location");
 
     TabletMutator tabletMutator = ample.mutateTablet(assignment.tablet);
     tabletMutator.putLocation(assignment.server, LocationType.FUTURE);
@@ -111,22 +117,23 @@ class ZooTabletStateStore implements TabletStateStore {
   }
 
   @Override
-  public void setLocation(Assignment assignment, TServerInstance prevLastLoc) {
+  public void setLocations(Collection<Assignment> assignments) throws DistributedStoreException {
+    if (assignments.size() != 1)
+      throw new IllegalArgumentException("There is only one root tablet");
+    Assignment assignment = assignments.iterator().next();
+    if (assignment.tablet.compareTo(RootTable.EXTENT) != 0)
+      throw new IllegalArgumentException("You can only store the root tablet location");
+
     TabletMutator tabletMutator = ample.mutateTablet(assignment.tablet);
     tabletMutator.putLocation(assignment.server, LocationType.CURRENT);
-    tabletMutator.putLocation(assignment.server, LocationType.LAST);
     tabletMutator.deleteLocation(assignment.server, LocationType.FUTURE);
-
-    if (prevLastLoc != null && !prevLastLoc.equals(assignment.server)) {
-      tabletMutator.deleteLocation(prevLastLoc, LocationType.LAST);
-    }
 
     tabletMutator.mutate();
   }
 
   @Override
   public void unassign(Collection<TabletLocationState> tablets,
-      Map<TServerInstance,List<Path>> logsForDeadServers) {
+      Map<TServerInstance,List<Path>> logsForDeadServers) throws DistributedStoreException {
     if (tablets.size() != 1)
       throw new IllegalArgumentException("There is only one root tablet");
     TabletLocationState tls = tablets.iterator().next();
@@ -155,7 +162,8 @@ class ZooTabletStateStore implements TabletStateStore {
 
   @Override
   public void suspend(Collection<TabletLocationState> tablets,
-      Map<TServerInstance,List<Path>> logsForDeadServers, long suspensionTimestamp) {
+      Map<TServerInstance,List<Path>> logsForDeadServers, long suspensionTimestamp)
+      throws DistributedStoreException {
     // No support for suspending root tablet.
     unassign(tablets, logsForDeadServers);
   }
